Moved route for application. Made Only user(type)
