+++ components/common/Project/MyApplication/index.tsx
@@ -31,7 +31,6 @@ function StatusApplications({
         group?.map((item) => (
           <BodyCard
             key={item.id}
-            username={item.user.username as string}
             applicationId={item.id}
             project={item.project}
             name={item.organization.meta.name}
